+++ src/pre_handle/entity/force_info_subbasin.h
@@ -10,10 +10,8 @@
  * forcing data(atmosphere info) of a subbasin node needed by simmulation
  */
 class force_info_subbasin {
-
 public:
     double air_tmp;
-
 };
 
 #endif //PNOHS_FORCE_INFO_SUBBASIN_H