+++ src/pre_handle/entity/geo_info_subbasin.h
@@ -16,5 +16,4 @@ public:
     double elev;
 };
 
-
 #endif //PNOHS_GEO_INFO_SUBBASIN_H