+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * **dependency:** add package.json for changelog ([eea21eb](https://git.gensh.me/HPCer/hydrology/pnohs/commits/eea21eb))
 * **init:** initial commit ([9a66a79](https://git.gensh.me/HPCer/hydrology/pnohs/commits/9a66a79))
 * **partition:** add partition interface ([d201fa7](https://git.gensh.me/HPCer/hydrology/pnohs/commits/d201fa7))
+* **partition:** read simulation dispatch results from partition file ([c1ccee9](https://git.gensh.me/HPCer/hydrology/pnohs/commits/c1ccee9))
 * **river-graph:** add river network graph ([2d56401](https://git.gensh.me/HPCer/hydrology/pnohs/commits/2d56401))
 * **simulation:** add simulation context ([342dace](https://git.gensh.me/HPCer/hydrology/pnohs/commits/342dace))